django_pos/requirements.txt
Django~=5.0.3
openpyxl==3.1.2
pillow==10.3.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
weasyprint==60.2
groq~=0.5.0
djangorestframework
whitenoise
psycopg2-binary
psycopg2
firebase-admin
intasend-python
drf-yasg